requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
yfinance>=0.2.0
